Moved glade UI into main
